requirements.txt
requests
python-dotenv
PyYAML
markdown
emoji